Refactored profile list views
