+++ src/meetix/lib/view/TopProfilesPage.dart
@@ -55,7 +55,6 @@ class _TopProfilesPageState extends State<TopProfilesPage> {
   Widget _buildList(BuildContext context, List<List<String>> profileIDs) {
     List<Widget> profiles = profileIDs.map((data) => _buildProfile(context, data)).toList();
     return ListView.separated(
-      shrinkWrap: true,
       padding: EdgeInsets.zero,
       itemCount: profiles.length,
       separatorBuilder: (context, index) => Divider(
